+++ src/main/java/it/cnr/ilc/lexo/service/Metadata.java
@@ -20,7 +20,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import org.apache.log4j.Level;
+import org.slf4j.event.Level;
 
 /** Common RDF metadata CRUD for supported application entities. */
 @Path("metadata")